+++ packages/kit/src/core/adapt/builder.js
@@ -5,6 +5,7 @@ import { pipeline } from 'stream';
 import { promisify } from 'util';
 import { copy, rimraf, mkdirp } from '../../utils/filesystem.js';
 import { generate_manifest } from '../generate_manifest/index.js';
+import { affects_path } from '../../utils/routing.js';
 
 const pipe = promisify(pipeline);
 
@@ -32,10 +33,11 @@ export function create_builder({ config, build_data, routes, prerendered, log })
 		async createEntries(fn) {
 			/** @type {import('types').RouteDefinition[]} */
 			const facades = routes.map((route) => {
+				/** @type {Set<import('types').HttpMethod>} */
 				const methods = new Set();
 
 				if (route.page) {
-					methods.add('SET');
+					methods.add('GET');
 				}
 
 				if (route.endpoint) {
@@ -46,11 +48,14 @@ export function create_builder({ config, build_data, routes, prerendered, log })
 
 				return {
 					id: route.id,
-					segments: route.id.split('/').map((segment) => ({
-						dynamic: segment.includes('['),
-						rest: segment.includes('[...'),
-						content: segment
-					})),
+					segments: route.id
+						.split('/')
+						.filter(affects_path)
+						.map((segment) => ({
+							dynamic: segment.includes('['),
+							rest: segment.includes('[...'),
+							content: segment
+						})),
 					pattern: route.pattern,
 					methods: Array.from(methods)
 				};
@@ -126,6 +131,10 @@ export function create_builder({ config, build_data, routes, prerendered, log })
 
 		getStaticDirectory() {
 			return config.kit.files.assets;
+		},
+
+		getAppPath() {
+			return build_data.app_path;
 		},
 
 		writeClient(dest) {