+++ packages/kit/src/runtime/server/data/index.js
@@ -15,7 +15,7 @@ import { DATA_SUFFIX } from '../../../constants.js';
  */
 export async function render_data(event, route, options, state) {
 	if (!route.page) {
-		// requesting /__data.js should fail for a +server.js
+		// requesting /__data.json should fail for a +server.js
 		return new Response(undefined, {
 			status: 404
 		});
@@ -25,10 +25,8 @@ export async function render_data(event, route, options, state) {
 		const node_ids = [...route.page.layouts, route.page.leaf];
 
 		const invalidated =
-			event.url.searchParams
-				.get('__invalid')
-				?.split('')
-				.map((x) => x === 'y') ?? node_ids.map(() => true);
+			event.request.headers.get('x-sveltekit-invalidated')?.split(',').map(Boolean) ??
+			node_ids.map(() => true);
 
 		let aborted = false;
 
@@ -37,9 +35,6 @@ export async function render_data(event, route, options, state) {
 			url.pathname.slice(0, -DATA_SUFFIX.length),
 			options.trailing_slash
 		);
-
-		url.searchParams.delete('__invalid');
-		url.searchParams.delete('__id');
 
 		const new_event = { ...event, url };
 
@@ -116,7 +111,7 @@ export async function render_data(event, route, options, state) {
 			nodes: nodes.slice(0, length)
 		};
 
-		return data_response(server_data);
+		return data_response(server_data, event);
 	} catch (e) {
 		const error = normalize_error(e);
 
@@ -127,10 +122,10 @@ export async function render_data(event, route, options, state) {
 				location: error.location
 			};
 
-			return data_response(server_data);
+			return data_response(server_data, event);
 		} else {
 			// TODO make it clearer that this was an unexpected error
-			return data_response(handle_error_and_jsonify(event, options, error));
+			return data_response(handle_error_and_jsonify(event, options, error), event);
 		}
 	}
 }