+++ packages/kit/src/utils/routing.js
@@ -1,4 +1,4 @@
-const param_pattern = /^(\.\.\.)?(\w+)(?:=(\w+))?$/;
+const param_pattern = /^(\[)?(\.\.\.)?(\w+)(?:=(\w+))?(\])?$/;
 
 /** @param {string} id */
 export function parse_route_id(id) {
@@ -22,57 +22,70 @@ export function parse_route_id(id) {
 						.map((segment, i, segments) => {
 							const decoded_segment = decodeURIComponent(segment);
 							// special case — /[...rest]/ could contain zero segments
-							const match = /^\[\.\.\.(\w+)(?:=(\w+))?\]$/.exec(decoded_segment);
-							if (match) {
-								names.push(match[1]);
-								types.push(match[2]);
+							const rest_match = /^\[\.\.\.(\w+)(?:=(\w+))?\]$/.exec(decoded_segment);
+							if (rest_match) {
+								names.push(rest_match[1]);
+								types.push(rest_match[2]);
 								return '(?:/(.*))?';
+							}
+							// special case — /[[optional]]/ could contain zero segments
+							const optional_match = /^\[\[(\w+)(?:=(\w+))?\]\]$/.exec(decoded_segment);
+							if (optional_match) {
+								names.push(optional_match[1]);
+								types.push(optional_match[2]);
+								return '(?:/([^/]+))?';
 							}
 
 							const is_last = i === segments.length - 1;
 
-							return (
-								decoded_segment &&
-								'/' +
-									decoded_segment
-										.split(/\[(.+?)\]/)
-										.map((content, i) => {
-											if (i % 2) {
-												const match = param_pattern.exec(content);
-												if (!match) {
-													throw new Error(
-														`Invalid param: ${content}. Params and matcher names can only have underscores and alphanumeric characters.`
-													);
-												}
+							if (!decoded_segment) {
+								return;
+							}
 
-												const [, rest, name, type] = match;
-												names.push(name);
-												types.push(type);
-												return rest ? '(.*?)' : '([^/]+?)';
-											}
+							const parts = decoded_segment.split(/\[(.+?)\](?!\])/);
+							const result = parts
+								.map((content, i) => {
+									if (i % 2) {
+										const match = param_pattern.exec(content);
+										if (!match) {
+											throw new Error(
+												`Invalid param: ${content}. Params and matcher names can only have underscores and alphanumeric characters.`
+											);
+										}
 
-											if (is_last && content.includes('.')) add_trailing_slash = false;
+										const [, optional, rest, name, type] = match;
+										// It's assumed that the following invalid route id cases are already checked
+										// - unbalanced brackets
+										// - optional param following rest param
 
-											return (
-												content // allow users to specify characters on the file system in an encoded manner
-													.normalize()
-													// We use [ and ] to denote parameters, so users must encode these on the file
-													// system to match against them. We don't decode all characters since others
-													// can already be epressed and so that '%' can be easily used directly in filenames
-													.replace(/%5[Bb]/g, '[')
-													.replace(/%5[Dd]/g, ']')
-													// '#', '/', and '?' can only appear in URL path segments in an encoded manner.
-													// They will not be touched by decodeURI so need to be encoded here, so
-													// that we can match against them.
-													// We skip '/' since you can't create a file with it on any OS
-													.replace(/#/g, '%23')
-													.replace(/\?/g, '%3F')
-													// escape characters that have special meaning in regex
-													.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')
-											); // TODO handle encoding
-										})
-										.join('')
-							);
+										names.push(name);
+										types.push(type);
+										return rest ? '(.*?)' : optional ? '([^/]*)?' : '([^/]+?)';
+									}
+
+									if (is_last && content.includes('.')) add_trailing_slash = false;
+
+									return (
+										content // allow users to specify characters on the file system in an encoded manner
+											.normalize()
+											// We use [ and ] to denote parameters, so users must encode these on the file
+											// system to match against them. We don't decode all characters since others
+											// can already be epressed and so that '%' can be easily used directly in filenames
+											.replace(/%5[Bb]/g, '[')
+											.replace(/%5[Dd]/g, ']')
+											// '#', '/', and '?' can only appear in URL path segments in an encoded manner.
+											// They will not be touched by decodeURI so need to be encoded here, so
+											// that we can match against them.
+											// We skip '/' since you can't create a file with it on any OS
+											.replace(/#/g, '%23')
+											.replace(/\?/g, '%3F')
+											// escape characters that have special meaning in regex
+											.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')
+									); // TODO handle encoding
+								})
+								.join('');
+
+							return '/' + result;
 						})
 						.join('')}${add_trailing_slash ? '/?' : ''}$`
 			  );
@@ -101,7 +114,7 @@ export function exec(match, names, types, matchers) {
 	for (let i = 0; i < names.length; i += 1) {
 		const name = names[i];
 		const type = types[i];
-		const value = match[i + 1] || '';
+		let value = match[i + 1] || '';
 
 		if (type) {
 			const matcher = matchers[type];